Allow different levels in each case branch.
The scrutinee's level can't exceed the ambient level
to ensure that the case expression doesn't depend on higher-level terms,
and the branches' levels can't exceed the scrutinee's level
to ensure that motives remain well typed in each branch.
The floating levels of the constructor arguments must be at the same level
as the constructor, which is the level of the branch for that constructor.